remove updates tab, and fix minor ui issues
