--- app/routes/my_updates.js
@@ -1,2 +1,0 @@
-import dashboard from "./my_dashboard";
-export default dashboard.extend({});
