+++ app/templates/my_dashboard/_my_dashboard_tabs.hbs
@@ -1,29 +1,19 @@
 <div class="row tabs tab-icons">
-  {{#link-to 'my_updates' tagName="div" classNames="small-4 updates tab columns text-center"}}
-    <i class="dashboard-menu">{{fa-icon 'bell-o'}}</i>
-  {{/link-to}}
-
-  {{#link-to 'my_orders' tagName="div" classNames="small-4 orders tab columns text-center"}}
+  {{#link-to 'my_orders' tagName="div" classNames="small-6 orders tab columns text-center"}}
     <i class="dashboard-menu">{{fa-icon 'shopping-cart'}}</i>
   {{/link-to}}
 
-  {{#link-to 'my_account' tagName="div" classNames="small-4 account tab columns text-center"}}
+  {{#link-to 'my_account' tagName="div" classNames="small-6 account tab columns text-center"}}
     <i class="dashboard-menu">{{fa-icon 'vcard'}}</i>
   {{/link-to}}
 </div>
 
 <div class="row tabs tab-name active-tab">
-  <div class="small-4 columns updates tab text-center">
-  </div>
-  {{#link-to 'my_updates' tagName="div" classNames="small-4 updates tab active-tab columns text-center"}}
-    {{t 'my_dashboard.updates'}}
-  {{/link-to}}
-
-  {{#link-to 'my_orders' tagName="div" classNames="small-4 orders tab active-tab columns text-center"}}
+  {{#link-to 'my_orders' tagName="div" classNames="small-6 orders tab active-tab columns text-center"}}
     {{t 'my_dashboard.orders'}}
   {{/link-to}}
 
-  {{#link-to 'my_account' tagName="div" classNames="small-4 account active-tab tab columns text-center"}}
+  {{#link-to 'my_account' tagName="div" classNames="small-6 account active-tab tab columns text-center"}}
     {{t 'my_dashboard.account.account_text'}}
   {{/link-to}}
 </div>
