--- app/templates/my_updates.hbs
@@ -1,9 +1,0 @@
-<div class="row my_dashboard">
-  <div class="title show-for-medium-up">{{t "my_dashboard.title"}}</div>
-  {{partial 'my_dashboard/my_dashboard_tabs'}}
-
-  <div class="small-12 columns content">
-    <div class="row my_updates">
-    </div>
-  </div>
-</div>
